pyproject.toml
[tool.poetry]
name = "moodly"
version = "0.1.0"
description = "Emotions Dairy Telegram-Bot Project"
authors = ["Artem Berdnikov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytelegrambotapi = "^4.26.0"
python-dotenv = "^1.0.1"
asyncpg-lite = "^0.3.1.3"
pydantic = "^2.10.5"
apscheduler = "^3.11.0"
pydantic-settings = "^2.7.1"
aiogram = "^3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
